+++ src/gridworld/qlearning.cpp
@@ -58,7 +58,7 @@ namespace gridworld_2D {
         return max_actions[std::rand() % max_actions.size()];
     }
 
-    int q_learning_agent::max_q(const state& cur) const {
+    float q_learning_agent::max_q(const state& cur) const {
         auto actions = grid_env::get_possible_actions(width_, height_, cur);
         float max = -std::numeric_limits<float>::infinity();
